2.86 done
All of the testing on the system from 2.81 to this exercises can be
considered part of 86-tests.scm in the tests directory, I think.

This is perhaps starting to turn into a hairy system now. Hopefully all
the work that was put into it this subsection will be used in the next
subsection as well. Off into the next section we go: Symbolic Algebra