2.92 and tests done -- see description
This is a naive implementation!

Alright, so in order to reduce the loss of hairs in my head and time
spent debugging, the system was greatly reduced here to deal with
numbers and polynomials. The concept here can be applied to the entirety
of the system for certain, but seeing how long 2.91 took me I thought it
would be a good idea to pass on that.

Need to get to chapter 3 asap!